feat: enable componentized scrubber
