+++ src/browser/transforms.js
@@ -309,11 +309,15 @@ function errorClass(stackInfo, guess, options) {
   }
 }
 
-function scrubPayload(item, options, callback) {
-  var scrubFields = options.scrubFields || [];
-  var scrubPaths = options.scrubPaths || [];
-  item.data = _.scrub(item.data, scrubFields, scrubPaths);
-  callback(null, item);
+function addScrubber(scrub) {
+  return function (item, options, callback) {
+    if (scrub) {
+      var scrubFields = options.scrubFields || [];
+      var scrubPaths = options.scrubPaths || [];
+      item.data = scrub(item.data, scrubFields, scrubPaths);
+    }
+    callback(null, item);
+  }
 }
 
 module.exports = {
@@ -325,5 +329,5 @@ module.exports = {
   addClientInfo: addClientInfo,
   addPluginInfo: addPluginInfo,
   addBody: addBody,
-  scrubPayload: scrubPayload
+  addScrubber: addScrubber
 };
