+++ src/utility/traverse.js
@@ -1,0 +1,38 @@
+var _ = require('../utility');
+
+function traverse(obj, func, seen) {
+  var k, v, i;
+  var isObj = _.isType(obj, 'object');
+  var isArray = _.isType(obj, 'array');
+  var keys = [];
+
+  if (isObj && seen.indexOf(obj) !== -1) {
+    return obj;
+  }
+  seen.push(obj);
+
+  if (isObj) {
+    for (k in obj) {
+      if (Object.prototype.hasOwnProperty.call(obj, k)) {
+        keys.push(k);
+      }
+    }
+  } else if (isArray) {
+    for (i = 0; i < obj.length; ++i) {
+      keys.push(i);
+    }
+  }
+
+  var result = isObj ? {} : [];
+  var same = true;
+  for (i = 0; i < keys.length; ++i) {
+    k = keys[i];
+    v = obj[k];
+    result[k] = func(k, v, seen);
+    same = same && result[k] === obj[k];
+  }
+
+  return (keys.length != 0 && !same) ? result : obj;
+}
+
+module.exports = traverse;
